Update docs/ so that the website can be rebuilt from source.
The sphinx-build function creates a directory called build/ that
contains the compiled html, assets, and a set of text files
corresponding to the source code. This is what was originally commited
to the repository.

Now, all that is being tracked is the source code in docs/source. To
rebuild the website, from the docs/ directory, run `make dirhtml`. This
will rebuild the website in docs/build/dirhtml.

Note that the notebooks are still out of date, and if they are ever
changed, the changes would need to be copied over somehow. Haven't
looked into whether it can be automated rather than done by hand.

Next step is to set of a GitLab runner to build the website when a
commit is pushed to master.