Update NCBI annotation script for accession numbers.
Should remain consistent with the previous version of the function,
though it defaults to assuming IDs used in the NCBI query are accession
numbers, not GI numbers.

It should be noted, though, that the efetch tool allows mixed GI and
acession numbers in queries, so the `--idtype` flag may be unnecessary.
(It defaults to 'acc', so the script can still run if the flag is
omitted.)