Use esummary to get taxonomy information for each sequence.
The issue now is that elink requires the non-redundant accession number
to be used instead of the GI number or the accession number directly
mapped to it. For lack of an easy way to obtain the former, I set the
annotation to use esummary instead. It returns taxonomy IDs.

It should be noted that the taxIDs returned by the reduntant (YP_...) vs
non-redundant (WP_...) numbers are different, but they seem to map to
the same lineage history from the handful of sequences I checked
manually.

For future reference, to use the elink tool, you need to specify
"dbfrom=protein", "id=<WP_...>", and "linkname=taxonomy_protein".

One disadvantage of using esummary is that the performance seems to be
worse (timeout errors, too many requests, etc.), even though the
documentation clains it can handle up to 200 IDs in a single request.
It could simply be that the servers are just slammed right now, but if
this problem persists, reducing id_blocksize seems to fix it.

I probably will look at this again. This solution is not by any means
ideal...