Switch shebang to python instead of python3.
Windows returns permissions errors (?) when python3 is set, but not when
python is set. Will have to leave it to users on other OSes to alias
their `python` executable to python3 if necessary.

(See the `set_alias()` function in tools/rcparams in the bmDCA
repository for an example on how.)