Scale basicICA learning rate based on the size of matrix to rotate.
Scaliing is 0.01 / M, where M is the length of the eigenvector. Seems to
help with overflow errors on larger alignments, but it is of course a
hack, not a good solution.

A better approach would be to test for super or sub-gaussianity and
chose the iterative step accordingly. It seems to work for some
matrices, but it's not clear yet that using tanh is an improvement over
the current logitistic function...